move benchmark to new scenario file
